collision detection done (contact face clipping pending)

+++ Engine/collision_detection.hpp
@@ -4,8 +4,13 @@
 #include "glm/vec3.hpp"
 
 struct Collision {
-	std::vector<glm::dvec3> collisionPointsA;
-	std::vector<glm::dvec3> collisionPointsB;
+	// in space of objectA
+	std::vector<glm::vec3> collisionPointsA;
+
+	// in space of objectB
+	std::vector<glm::vec3> collisionPointsB;
+
+	// A to B
 	glm::vec3 collisionNormal;
 };
 class Gameobject;
